ODK-25: Ability to review already submitted surveys (read-only) - part 1
Added posibility to view the submitted surveys in a read-only mode. The exact same prompts are
displayed to the user but with all inputs, textareas, button and selects blocked. Custom forms
within prompts are blocked by disabling touch interaction witch prevents any modification to
the survey but it keeps the ability to scroll the view content.

TODO:
* connect the read-only view of the survey with the instant summary view from ODK-24